Adds ecr client auth options
